fix: boardPost, reviewPost 사용자 검증 추가

+++ src/main/java/com/tave/camchelin/domain/board_posts/service/BoardPostService.java
@@ -9,6 +9,7 @@ import com.tave.camchelin.domain.review_posts.dto.ReviewPostDto;
 import com.tave.camchelin.domain.users.entity.User;
 import com.tave.camchelin.domain.users.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,9 +39,9 @@ public class BoardPostService {
     }
 
     @Transactional
-    public BoardPostDto writeBoardPost(BoardPostDto boardPostDto) {
+    public BoardPostDto writeBoardPost(Long userId, BoardPostDto boardPostDto) {
         // User와 Community를 찾아 연관 관계를 설정
-        User user = userRepository.findById(boardPostDto.getUser().getId())
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("유저 정보를 찾을 수 없습니다."));
         Community community = communityRepository.findByName(boardPostDto.getCommunity().getName())
                 .orElseThrow(() -> new IllegalArgumentException("커뮤니티 정보를 찾을 수 없습니다."));
@@ -58,9 +59,13 @@ public class BoardPostService {
     }
 
     @Transactional
-    public void editBoardPost(Long boardPostId, BoardPostDto boardPostDto) {
+    public void editBoardPost(Long userId, Long boardPostId, BoardPostDto boardPostDto) {
         BoardPost boardPost = boardPostRepository.findById(boardPostId)
                 .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+
+        if (!boardPost.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("권한이 없습니다."); // Spring Security 예외 사용
+        }
 
         // 수정 가능한 필드만 업데이트
         boardPost.edit(boardPostDto.getTitle(), boardPostDto.getContent());
@@ -69,9 +74,14 @@ public class BoardPostService {
     }
 
     @Transactional
-    public void deleteBoardPost(Long boardPostId) {
+    public void deleteBoardPost(Long userId, Long boardPostId) {
         BoardPost boardPost = boardPostRepository.findById(boardPostId)
                 .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+
+        if (!boardPost.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("권한이 없습니다."); // Spring Security 예외 사용
+        }
+
         boardPostRepository.delete(boardPost);
     }
 }
